Add files via upload
This repository contains all the files required to run the Physics Quiz and Certificate Generation Platform, a digital solution designed to replace traditional paper-based quizzes. The platform includes a secure student database, automated grading, and a certificate system that awards Bronze, Silver, or Gold certificates based on performance.

Additionally, a Test Plan file is included to guide you through the functionality of each section and feature of the platform, providing a clear overview of how the system operates. This project demonstrates practical skills in software development, database management and automation.